app/requirements.txt
streamlit
faiss-cpu
sentence-transformers
# は重いので、scikit-learnでOK
# scikit-learn
pypdf
requests
numpy